feat: 최신 백엔드 스펙에 맞춘 D-Balance 앱 완전 재구현
✅ 패키지 설치 및 설정:
- flutter_naver_login 추가
- kakao_flutter_sdk_user로 업데이트
- 모든 소셜 로그인 패키지 최신화

✅ Core 네트워킹 구현:
- DioClient: BaseOptions + 인터셉터 설정
- AuthInterceptor: X-Session-Token 자동 첨부, 401 refresh 처리
- SecureStorage: 토큰 저장, 복구 코드, 사용자 정보 관리

✅ 데이터 모델 및 Repository:
- User, SessionInfo, Room, Round, Participant, Expense 모델
- json_serializable로 자동 생성
- AuthRepository, SessionRepository, RoomsRepository, ItemsRepository
- 최신 백엔드 API 스펙 반영

✅ Provider 구성:
- sessionBootstrapProvider: 게스트 세션 자동 생성
- authStateProvider: guest/signedIn 상태 관리
- roomsRecentProvider: 최근 방 목록 (게스트 3개 제한)
- roomDetailProvider, roundsProvider, expensesProvider

✅ 화면 구현:
- SplashScreen: 세션 부트스트랩 후 홈으로 이동
- LoginScreen: 이메일/비밀번호 + 소셜 로그인 (Google/Apple/Kakao/Naver)
- HomeScreen: 최근 방 목록, 게스트 배너, 방 생성
- RoomDetailScreen: 방 정보, 라운드/참가자/지출 탭

✅ 라우팅:
- Go Router 설정
- 인증 상태별 리다이렉트 로직
- 에러 페이지 처리

✅ 주요 기능:
- 게스트 세션 자동 생성 및 저장
- 모든 API 요청에 X-Session-Token 자동 첨부
- 로그인 시 Authorization Bearer 토큰 첨부
- 401 에러 시 자동 토큰 갱신 및 재시도
- 게스트/로그인 상태별 UI 분기

⚠️ 현재 상태:
- 앱 실행 시 크래시 발생 (Lost connection to device)
- 디버깅 필요

+++ lib/shared/models/expense.dart
@@ -8,7 +8,7 @@ class Expense {
   final int roomId;
   final int roundId;
   final int payerParticipantId;
-  final double totalAmount;
+  final int totalAmount;
   final String? currency;
   final String? memo;
   final List<ExpenseItem> items;
@@ -33,7 +33,7 @@ class ExpenseItem {
   final int id;
   final int expenseId;
   final String category;
-  final double amount;
+  final int amount;
   final List<ParticipantOverride>? participantsOverride;
 
   const ExpenseItem({
